move float -> double

+++ core/include/CLInterval.h
@@ -17,11 +17,11 @@ class CLInterval {
  public:
   void print() const;
 
-  float pvalue = -1.;             // pvalue corresponding to this interval
-  float pvalueAtCentral = -1.;    // pvalue at the central value
-  float min = -1.;                // lower interval border
-  float max = -1.;                // upper interval border
-  float central = -1.;            // central value
+  double pvalue = -1.;            // pvalue corresponding to this interval
+  double pvalueAtCentral = -1.;   // pvalue at the central value
+  double min = -1.;               // lower interval border
+  double max = -1.;               // upper interval border
+  double central = -1.;           // central value
   bool minclosed = false;         // true if the interval was not closed by limited scan range
   bool maxclosed = false;         // true if the interval was not closed by limited scan range
   TString minmethod = "n/a";      // details on the algorithm that found this interval
