+++ core/include/ToyTree.h
@@ -32,11 +32,11 @@ class ToyTree {
   Long64_t GetEntries() const;
   void GetEntry(Long64_t i);
   inline TString getName() const { return name; };
-  float getScanpointMin();
-  float getScanpointMax();
+  double getScanpointMin();
+  double getScanpointMax();
   int getScanpointN();
-  float getScanpointyMin();
-  float getScanpointyMax();
+  double getScanpointyMin();
+  double getScanpointyMax();
   int getScanpointyN();
   TTree* getTree() { return t; };
   bool isWsVarAngle(TString var) const;
@@ -55,47 +55,48 @@ class ToyTree {
   void setStoreGlob(bool flag) { this->storeTh = flag; };
   void storeParsGau(const RooArgSet globalConstraintMeans);
 
-  float scanpoint;              ///< the scanpoint for 1D scans, or the x scanpoint for 2D scans
-  float scanpointy = 0.;        ///< the y scanpoint for 2D scans
-  float chi2min = 0.;           ///< the chi2 of the fit with var fixed to scan point
-  float chi2minGlobal = 0.;     ///< the chi2 of the free fit
-  float chi2minBkg = 0.;        ///< the chi2 of the fit of the bkg hypothesis (for CLs method)
-  float chi2minToy = 0.;        ///< the chi2 of the fit to the toy with var fixed to scan point
-  float chi2minGlobalToy = 0.;  ///< the chi2 of the free fit to the toy
-  float chi2minBkgToy =
+  double scanpoint;              ///< the scanpoint for 1D scans, or the x scanpoint for 2D scans
+  double scanpointy = 0.;        ///< the y scanpoint for 2D scans
+  double chi2min = 0.;           ///< the chi2 of the fit with var fixed to scan point
+  double chi2minGlobal = 0.;     ///< the chi2 of the free fit
+  double chi2minBkg = 0.;        ///< the chi2 of the fit of the bkg hypothesis (for CLs method)
+  double chi2minToy = 0.;        ///< the chi2 of the fit to the toy with var fixed to scan point
+  double chi2minGlobalToy = 0.;  ///< the chi2 of the free fit to the toy
+  double chi2minBkgToy =
       0.;  ///< the chi2 of the fit of the hypothesis value to the bkg toy distribution (for CLs method)
-  float chi2minGlobalBkgToy = 0.;  ///< the chi2 of the free fit to the bkg only toys
-  float chi2minBkgBkgToy = 0.;     ///< the chi2 of the bkg fit to the bkg only toys
-  float scanbest = 0.;             ///< an alias to the free fit value of the scan variable
-  float scanbesty = 0.;            ///< an alias to the free fit value of the scan y variable in 2D scans
-  float scanbestBkg = 0.;  ///< an alias to the free fit value of the scan variable on the bkg only toy (for CLs method)
-  float scanbestBkgfitBkg =
+  double chi2minGlobalBkgToy = 0.;  ///< the chi2 of the free fit to the bkg only toys
+  double chi2minBkgBkgToy = 0.;     ///< the chi2 of the bkg fit to the bkg only toys
+  double scanbest = 0.;             ///< an alias to the free fit value of the scan variable
+  double scanbesty = 0.;            ///< an alias to the free fit value of the scan y variable in 2D scans
+  double scanbestBkg =
+      0.;  ///< an alias to the free fit value of the scan variable on the bkg only toy (for CLs method)
+  double scanbestBkgfitBkg =
       0.;  ///< an alias to the fit value of the scan variable of the bkg fit on the bkg only toy (for CLs method)
-  float nrun = 0.;    ///< an ID to distinguish different runs, i.e. batch jobs
-  float ntoy = 0.;    ///< an ID to distinguish different toys
-  float npoint = 0.;  ///< an ID to distinguish different scan point
-  float id = 0.;      ///< an ID to distinguish different conditions, e.g. different toys in a coverage test
-  float statusFree = -5.;
-  float covQualFree = -2.;
-  float statusScan = -5.;
-  float covQualScan = -2.;
-  float statusFreeBkg = -5.;
-  float covQualFreeBkg = -2.;
-  float statusScanBkg = -5.;
-  float covQualScanBkg = -2.;
-  float statusBkgBkg = -5.;
-  float covQualBkgBkg = -2.;
-  float statusScanData = -5.;
-  float covQualScanData = -2.;
+  double nrun = 0.;    ///< an ID to distinguish different runs, i.e. batch jobs
+  double ntoy = 0.;    ///< an ID to distinguish different toys
+  double npoint = 0.;  ///< an ID to distinguish different scan point
+  double id = 0.;      ///< an ID to distinguish different conditions, e.g. different toys in a coverage test
+  double statusFree = -5.;
+  double covQualFree = -2.;
+  double statusScan = -5.;
+  double covQualScan = -2.;
+  double statusFreeBkg = -5.;
+  double covQualFreeBkg = -2.;
+  double statusScanBkg = -5.;
+  double covQualScanBkg = -2.;
+  double statusBkgBkg = -5.;
+  double covQualBkgBkg = -2.;
+  double statusScanData = -5.;
+  double covQualScanData = -2.;
   int bestIndexScanData = 0;
-  float nBergerBoos = 0.;
-  float BergerBoos_id = 0.;
-  float genericProbPValue = 0.;
-  float statusFreePDF = -5.;
-  float statusScanPDF = -5.;
-  float chi2minToyPDF = 0.;
-  float chi2minGlobalToyPDF = 0.;
-  float chi2minBkgToyPDF = 0.;
+  double nBergerBoos = 0.;
+  double BergerBoos_id = 0.;
+  double genericProbPValue = 0.;
+  double statusFreePDF = -5.;
+  double statusScanPDF = -5.;
+  double chi2minToyPDF = 0.;
+  double chi2minGlobalToyPDF = 0.;
+  double chi2minBkgToyPDF = 0.;
   TTree* t = nullptr;  ///< the tree
 
  private:
@@ -110,19 +111,19 @@ class ToyTree {
   TString thName = "";             ///< set name of theory parameters, derived from name
   TString globName;                ///< set name of explicit set of global observables
 
-  std::map<std::string, float> parametersScan;  ///< fit result of the scan fit
-  std::map<std::string, float> parametersFree;  ///< fit result of the free fit
-  std::map<std::string, float> parametersPll;   ///< parameters of the profile likelihood curve of the data
-  std::map<std::string, float> observables;     ///< values of the observables
-  std::map<std::string, float> theory;          ///< theory parameters (=observables at profile likelihood points)
-  std::map<TString, float> constraintMeans;     ///< values of global observables
+  std::map<std::string, double> parametersScan;  ///< fit result of the scan fit
+  std::map<std::string, double> parametersFree;  ///< fit result of the free fit
+  std::map<std::string, double> parametersPll;   ///< parameters of the profile likelihood curve of the data
+  std::map<std::string, double> observables;     ///< values of the observables
+  std::map<std::string, double> theory;          ///< theory parameters (=observables at profile likelihood points)
+  std::map<TString, double> constraintMeans;     ///< values of global observables
 
-  float scanpointMin = 0.;   ///< minimum of the scanpoint, computed by computeMinMaxN().
-  float scanpointMax = 0.;   ///< maximum of the scanpoint, computed by computeMinMaxN().
-  int scanpointN = -1.;      ///< number of different values of the scanpoint, computed by computeMinMaxN().
-  float scanpointyMin = 0.;  ///< minimum of the scanpointy, computed by computeMinMaxN().
-  float scanpointyMax = 0.;  ///< maximum of the scanpointy, computed by computeMinMaxN().
-  int scanpointyN = -1.;     ///< number of different values of the scanpointy, computed by computeMinMaxN().
+  double scanpointMin = 0.;   ///< minimum of the scanpoint, computed by computeMinMaxN().
+  double scanpointMax = 0.;   ///< maximum of the scanpoint, computed by computeMinMaxN().
+  int scanpointN = -1.;       ///< number of different values of the scanpoint, computed by computeMinMaxN().
+  double scanpointyMin = 0.;  ///< minimum of the scanpointy, computed by computeMinMaxN().
+  double scanpointyMax = 0.;  ///< maximum of the scanpointy, computed by computeMinMaxN().
+  int scanpointyN = -1.;      ///< number of different values of the scanpointy, computed by computeMinMaxN().
 
   bool storeObs;   ///< Boolean flag to control storing ToyTree observables, can't store these for DatasetsScans
   bool storeTh;    ///< Boolean flag to control storing ToyTree theory parameters. Not needed in DatasetsScans
